𝙲𝚈𝙱𝙴𝚁 𝙱𝙾𝚃  𝕟𝕖𝕨 𝕦𝕡𝕕𝕒𝕥𝕖
Update and bug fixed

+++ plugins/updater.js
@@ -1,9 +1,9 @@
-/* Copyright (C) 2021 Aqua Snake.
+/* Copyright (C) 2021 Cyber Bot.
 
 Licensed under the  GPL-3.0 License;
 you may not use this file except in compliance with the License.
 
-Cyber Army Bot  - Aqua-Snake
+Cyber Bot - Aqua Snake
 */
 
 const simpleGit = require('simple-git');
@@ -15,13 +15,12 @@ const exec = require('child_process').exec;
 const Heroku = require('heroku-client');
 const { PassThrough } = require('stream');
 const heroku = new Heroku({ token: Config.HEROKU.API_KEY })
-const CBotStack = require('CBot-npm');
+
 const Language = require('../language');
 const Lang = Language.getString('updater');
 
 
-CBot.addCommand({pattern: 'update$', fromMe: true, desc: Lang.UPDATER_DESC}, (async (message, match) => {
-    await WhatsCBotStack.startwhatsCBot()
+CBot.applyCMD({pattern: 'update$', fromMe: true, desc: Lang.UPDATER_DESC, dontAddCommandList: true,  deleteCommand: false}, (async (message, match) => {
     await git.fetch();
     var commits = await git.log([Config.BRANCH + '..origin/' + Config.BRANCH]);
     if (commits.total === 0) {
@@ -33,7 +32,7 @@ CBot.addCommand({pattern: 'update$', fromMe: true, desc: Lang.UPDATER_DESC}, (as
         var degisiklikler = Lang.NEW_UPDATE;
         commits['all'].map(
             (commit) => {
-                degisiklikler += '▫️ [' + commit.date.substring(0, 10) + ']: ' + commit.message + ' <' + commit.author_name + '>\n';
+                degisiklikler += '🆕 [' + commit.date.substring(0, 10) + ']: ' + commit.message + ' \n <' + commit.author_name + '>\n\n';
             }
         );
         
@@ -43,18 +42,9 @@ CBot.addCommand({pattern: 'update$', fromMe: true, desc: Lang.UPDATER_DESC}, (as
         ); 
     }
 }));
-var Action = ''
-if (Config.LANG == 'TR') Action = '*CBot Halihazırda Güncelleniyor!*'
-if (Config.LANG == 'AZ') Action = '*CBot Hal -hazırda Yenilənir!*'
-if (Config.LANG == 'EN') Action = '*CBot Currently Updating!*'
-if (Config.LANG == 'RU') Action = '*CBot сейчас обновляется!*'
-if (Config.LANG == 'ES') Action = '*CBot Actualizando actualmente!*'
-if (Config.LANG == 'PT') Action = '*O CBot está sendo atualizado no momento!*'
-if (Config.LANG == 'ML') Action = '*CBot നിലവിൽ അപ്ഡേറ്റ് ചെയ്യുന്നു!*'
-if (Config.LANG == 'HI') Action = '*CBot वर्तमान में अपडेट हो रहा है!*'
-if (Config.LANG == 'ID') Action = '*CBot Saat Ini Memperbarui!*'
-CBot.addCommand({pattern: 'update now$', fromMe: true, desc: Lang.UPDATE_NOW_DESC}, (async (message, match) => {
-    await CBotStack.startCBot()
+
+
+CBot.applyCMD({pattern: 'update now$', fromMe: true, desc: Lang.UPDATE_NOW_DESC, dontAddCommandList: true}, (async (message, match) => {
     await git.fetch();
     var commits = await git.log([Config.BRANCH + '..origin/' + Config.BRANCH]);
     if (commits.total === 0) {
@@ -63,8 +53,6 @@ CBot.addCommand({pattern: 'update now$', fromMe: true, desc: Lang.UPDATE_NOW_DES
             Lang.UPDATE, MessageType.text
         );    
     } else {
-        var on_progress = false
-        if (on_progress) return await message.client.sendMessage(message.jid,Action,MessageType.text)
         var guncelleme = await message.reply(Lang.UPDATING);
         if (Config.HEROKU.HEROKU) {
             try {
@@ -83,7 +71,7 @@ CBot.addCommand({pattern: 'update now$', fromMe: true, desc: Lang.UPDATE_NOW_DES
             var git_url = app.git_url.replace(
                 "https://", "https://api:" + Config.HEROKU.API_KEY + "@"
             )
-            on_progress = true
+            
             try {
                 await git.addRemote('heroku', git_url);
             } catch { console.log('heroku remote ekli'); }
@@ -102,7 +90,62 @@ CBot.addCommand({pattern: 'update now$', fromMe: true, desc: Lang.UPDATE_NOW_DES
                     exec('npm install').stderr.pipe(process.stderr);
                 } else if (err) {
                     await message.client.sendMessage(
-                        message.jid,'*❌ Güncelleme başarısız oldu!*\n*Hata:* ```' + err + '```', MessageType.text);
+                        message.jid,'*❌ Update faield!*\n*Error:* ```' + err + '```', MessageType.text);
+                }
+            }));
+            await guncelleme.delete();
+        }
+    }
+}));
+
+
+CBot.applyCMD({pattern: 'verify$', fromMe: true, desc: Lang.VERIFY_DESC, dontAddCommandList: true}, (async (message, match) => {
+    await git.fetch();
+    var commits = await git.log([Config.BRANCH + '..origin/' + Config.BRANCH]);
+    if (commits.total === 0) {
+        return await message.client.sendMessage(
+            message.jid,
+            Lang.VERIFY, MessageType.text
+        );    
+    } else {
+        var guncelleme = await message.reply(Lang.VERIFYING);
+        if (Config.HEROKU.HEROKU) {
+            try {
+                var app = await heroku.get('/apps/' + Config.HEROKU.APP_NAME)
+            } catch {
+                await message.client.sendMessage(
+                    message.jid,Lang.INVALID_HEROKU, MessageType.text);
+                await new Promise(r => setTimeout(r, 1000));
+                return await message.client.sendMessage(
+                    message.jid,Lang.IN_AF, MessageType.text);
+            }
+
+            git.fetch('upstream', Config.BRANCH);
+            git.reset('hard', ['FETCH_HEAD']);
+
+            var git_url = app.git_url.replace(
+                "https://", "https://api:" + Config.HEROKU.API_KEY + "@"
+            )
+            
+            try {
+                await git.addRemote('heroku', git_url);
+            } catch { console.log('heroku remote ekli'); }
+            await git.push('heroku', Config.BRANCH);
+
+            await message.client.sendMessage(
+                message.jid,Lang.VERIFIED, MessageType.text);
+
+            await message.sendMessage(Lang.AFTER_UPDATE);
+            
+        } else {
+            git.pull((async (err, update) => {
+                if(update && update.summary.changes) {
+                    await message.client.sendMessage(
+                        message.jid,Lang.VERIFIED_LOCAL, MessageType.text);
+                    exec('npm install').stderr.pipe(process.stderr);
+                } else if (err) {
+                    await message.client.sendMessage(
+                        message.jid,'*❌ Verification faield!*\n*Error:* ```' + err + '```', MessageType.text);
                 }
             }));
             await guncelleme.delete();
